+++ CHANGELOG.rst
@@ -27,6 +27,9 @@ Addded
 
 - Stratmann solver for excited state, including range separated calculations
 
+- Rational function geometry optimization driver
+
+
 Changed
 -------
 
@@ -34,6 +37,9 @@ Changed
   <https://fpm.fortran-lang.org/>`_ preferred structure.
 
 - Updated parser version to 10.
+
+- New geometry optimization driver
+
 
 Fixed
 -----
@@ -49,6 +55,11 @@ Fixed
 - Updated version of dftd3-lib fixes a numerical typo in the Ba parameters
 
 - D3(BJ)-ATM calculator was not being passed the exponent for ATM zero damping calculations
+
+
+- LBFGS implementation fixed in new geometry optimization driver
+
+- Coupled cartesian and lattice parameter optimization
 
 
 21.1 (2021-05-12)